requirements.txt
pkg_resources==0.0.0
APScheduler==3.6.3
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.6.0
cachetools==4.2.2
certifi==2022.6.15
click==8.1.3
decorator==5.1.1
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
stack-data==0.4.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
wcwidth==0.2.5
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.12
anyio==3.6.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
python-dotenv==0.20.0
python-telegram-bot==20.0a2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
